requirements.txt
sentence-transformers==2.2.2
huggingface-hub==0.14.1
transformers==4.29.2
torch==2.0.1
pinecone-client==3.2.0
langchain==0.1.16
langchain-community==0.0.34
langchain-pinecone==0.0.3
langchain-experimental==0.0.11
pypdf==3.17.4
python-dotenv==1.0.1
langchain-google-genai==0.0.9
streamlit
